requirements.txt
asgiref==3.2.10
attrs==20.2.0
awsebcli==3.19.0
bcrypt==3.2.0
blessed==1.17.10
botocore==1.17.63
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
Django==3.1.2
django-crispy-forms==1.9.2
django-summernote==0.8.11.6
django-taggit==1.3.0
djangorestframework==3.12.2
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
importlib-metadata==2.0.0
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.2.2
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.3
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.3.0
asgiref==3.2.10
attrs==20.2.0
awsebcli==3.19.0
bcrypt==3.2.0
blessed==1.17.10
botocore==1.17.63
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
Django==3.1.2
django-crispy-forms==1.9.2
django-summernote==0.8.11.6
django-taggit==1.3.0
djangorestframework==3.12.2
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
importlib-metadata==2.0.0
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.2.2
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.3
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.3.0
